==> R/visualisation/04_fossil_palaeolat.R <==
# ------------------------------------------------------------------------- #
# Purpose: Plot fossil palaeolat range
# Author(s): Lucas Buffan & Lewis A. Jones
# Email: Lucas.L.Buffan@gmail.com; LewisAlan.Jones@uvigo.es
# Load libraries ----------------------------------------------------------
library(ggplot2)
library(ggpubr)
library(matrixStats)
library(Hmisc)
# Extract palaeolatitude from rotated occurrences -----------------------------
extract_palaeolat <- function(occ_idx, ds, lat.cols){
  sub <- as.character(ds[occ_idx, lat.cols])
  sub <- suppressWarnings(unlist(lapply(X = sub, FUN = as.numeric)))
  if(length(which(is.na(sub) == FALSE)) == 0){
    return(NA)
  }
  else{
    return(sub[which(is.na(sub) == FALSE)])
  }
}

lat.cols <- unlist(lapply(X = seq(from = 5, to = 235, by = 10),
                          FUN = function(x){return(paste0("lat_", x))}))
# Coral reefs -----------------------------------------------------------------
# List files
files <- list.files("./python/data/reef_palaeorot/", full.names = TRUE)
# Coral Reefs
df <- data.frame(time = read.csv(files[1])$mid_ma,
                 min = NA,
                 median = NA,
                 max = NA)
for (i in files) {
  df_mod <- read.csv(i)
  palaeolat <- unlist(lapply(X = 1:nrow(df_mod), FUN = extract_palaeolat, ds = df_mod, lat.cols = lat.cols))
  file <- strsplit(i, split = "/")[[1]][length(strsplit(i, split = "/")[[1]])]
  df <- cbind.data.frame(df, palaeolat)
}
colnames(df)[5:9] <- sub("_.*", "", sub(".*/", "", files))
# Get stats
df$min <- rowMins(as.matrix.data.frame(df[5:9]), na.rm = TRUE)
df$median <- rowMedians(as.matrix.data.frame(df[5:9]), na.rm = TRUE)
df$max <- rowMaxs(as.matrix.data.frame(df[5:9]), na.rm = TRUE)
# Number of models - NAs
df$count <- 5 - rowSums2(is.na(df[5:9]), na.rm = TRUE)

# Remove Inf data for plotting
df <- df[-which(is.infinite(df$max) == TRUE), ]

# Save dataframe for next plot
saveRDS(df, "./data/fossil_palaeocoordinates/reef_palaeoccordinates_5_models.RDS")

#Correlation time~range
TIME <- sort(unique(df$time)) #time intervals covered by df
average_range <- c()
for(t in TIME){
  ind <- which(df$time == t)
  average_range <- c(average_range, mean(df$max[ind] - df$min[ind]))
}
res <- Hmisc::rcorr(x = TIME, y = average_range, type = "pearson")
message(paste0("Corals time~uncertainty correlation coefficient (r): ", round(res$r[1,2], digits = 2)))
message(paste0("Corals time~uncertainty correlation p-value (p): ", round(res$P[1,2], digits = 7)))

# Plot
p1 <- ggplot(data = df[, 1:4], aes(x = time, y = median)) +
  geom_errorbar(aes(ymin = min, ymax = max), colour = "#006837") +
  geom_point(shape = 21, size = log(df$count), colour = "black", fill = "#006837", alpha = 0.7) +
  scale_y_continuous(limits = c(-90, 90),
                     labels = seq(-90, 90, 30),
                     breaks = seq(-90, 90, 30)) + 
  scale_x_reverse(limits = c(235, 0)) +
  labs(title = "Coral Reefs",
       x = "Time (Ma)",
       y = "Palaeolatitude (\u00B0)") +
  theme(plot.margin = margin(5, 10, 5, 10, "mm"),
        axis.title.x = element_text(size = 14),
        axis.title.y = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.background = element_blank(),
        legend.title = element_text(size = 10),
        legend.text = element_text(size = 10),
        legend.key.size = unit(0.5, "cm"),
        legend.position = c(0.93, 0.15),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), 
        panel.background = element_blank(),
        panel.border = element_rect(colour = "black", fill = NA, linewidth = 0.5)) +
  deeptime::coord_geo(pos = "bottom", fill = "grey95", height = unit(1.5, "line"))
# Crocs ------------------------------------------------------------------
# List files
files <- list.files("./python/data/croc_palaeorot/", full.names = TRUE)
# Crocs
df <- data.frame(time = read.csv(files[1])$mid_ma,
                 min = NA,
                 median = NA,
                 max = NA)
for (i in files) {
  df_mod <- read.csv(i)
  palaeolat <- unlist(lapply(X = 1:nrow(df_mod), FUN = extract_palaeolat, ds = df_mod, lat.cols = lat.cols))
  df <- cbind.data.frame(df, palaeolat)
}
colnames(df)[5:9] <- sub("_.*", "", sub(".*/", "", files))
# Get stats
df$min <- rowMins(as.matrix.data.frame(df[5:9]), na.rm = TRUE)
df$median <- rowMedians(as.matrix.data.frame(df[5:9]), na.rm = TRUE)
df$max <- rowMaxs(as.matrix.data.frame(df[5:9]), na.rm = TRUE)
# Number of models - NAs
df$count <- 5 - rowSums2(is.na(df[5:9]), na.rm = TRUE)

# Save dataframe for next plot
saveRDS(df, "./data/fossil_palaeocoordinates/croc_palaeoccordinates_5_models.RDS")

#Correlation time~range
TIME <- sort(unique(df$time)) #time intervals covered by df
average_range <- c()
for(t in TIME){
  ind <- which(df$time == t)
  average_range <- c(average_range, mean(df$max[ind] - df$min[ind]))
}
res <- Hmisc::rcorr(x = TIME, y = average_range, type = "pearson")

message(paste0("Crocs time~uncertainty correlation coefficient (r): ", round(res$r[1,2], digits = 2)))
message(paste0("Crocs time~uncertainty correlation p-value (p): ", round(res$P[1,2], digits = 7)))

# Plot
p2 <- ggplot(data = df[, 1:4], aes(x = time, y = median)) +
  geom_errorbar(aes(ymin = min, ymax = max), colour = "#7a0177") +
  geom_point(shape = 21, size = log(df$count), colour = "black", fill = "#7a0177", alpha = 0.7) +
  scale_y_continuous(limits = c(-90, 90),
                     labels = seq(-90, 90, 30),
                     breaks = seq(-90, 90, 30)) + 
  scale_x_reverse(limits = c(235, 0)) +
  labs(title = "Crocodylomorphs",
       x = "Time (Ma)",
       y = "Palaeolatitude (\u00B0)") +
  theme(plot.margin = margin(5, 10, 5, 10, "mm"),
        axis.title.x = element_text(size = 14),
        axis.title.y = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.background = element_blank(),
        legend.title = element_text(size = 10),
        legend.text = element_text(size = 10),
        legend.key.size = unit(0.5, "cm"),
        legend.position = c(0.93, 0.15),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), 
        panel.background = element_blank(),
        panel.border = element_rect(colour = "black", fill = NA, linewidth = 0.5)) +
  deeptime::coord_geo(pos = "bottom", fill = "grey95", height = unit(1.5, "line"))

# Combine plots -----------------------------------------------------------
# Arrange plot
p <- ggarrange(p1, p2, ncol = 1, nrow = 2, labels = c("(a)", "(b)"),
               font.label = list(size = 20))
# Save plot
ggsave(filename = "./figures/lat_range.png",
       height = 250,
       width = 210,
       units = "mm",
       dpi = 600)
